Rename build settings and opt for that instead
