+++ config/webpack.config.js
@@ -165,7 +165,6 @@ const configGenerator = (options, apps) => {
           WEB_PORT: (process.env.WEB_PORT || 3333),
           API_URL: process.env.API_URL ? JSON.stringify(process.env.API_URL) : null,
           BASE_URL: process.env.BASE_URL ? JSON.stringify(process.env.BASE_URL) : null,
-          BRAND_CONSOLIDATION_ENABLED: JSON.stringify(process.env.BRAND_CONSOLIDATION_ENABLED)
         }
       }),
